requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.22,<1.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.24,<1.0
pydantic>=2.0,<3.0
python-multipart>=0.0.9,<1.0
